Move custom AudioLoadResultHandler implementation to own class
